Separate error pages into their own components
Also add pages for new error codes
