+++ src/components/ErrorPages/404/404.module.scss
@@ -1,0 +1,22 @@
+.content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  flex-grow: 1;
+  gap: 1.5rem;
+  text-align: center;
+}
+
+.header {
+  font-size: 16rem;
+  color: var(--border-color);
+}
+
+.actionLink {
+  text-decoration: underline;
+}
+
+.homeLink {
+  font-size: 1.5rem;
+}
